# src/inventory/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple rest api for category",
        "title": "User API documentation",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/product": {
            "get": {
                "description": "Get the list all the products",
                "tags": [
                    "Product"
                ],
                "summary": "Get all the products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Product"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "description": "Create new product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Create new product based on parameters",
                "parameters": [
                    {
                        "description": "Product Data",
                        "name": "Product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete all the existing products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Delete all the existing products",
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/product/delete/{id}": {
            "delete": {
                "description": "Delete existing product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Delete existing product based on id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/product/update/{id}": {
            "patch": {
                "description": "Update existing Product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Update existing product based on parameters",
                "parameters": [
                    {
                        "description": "Product Data",
                        "name": "Product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/product/{id}": {
            "get": {
                "description": "get product by ID",
                "tags": [
                    "Product"
                ],
                "summary": "Get one product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.NullTime": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.Product": {
            "type": "object",
            "required": [
                "category_id",
                "cost_price",
                "product_name",
                "quantity"
            ],
            "properties": {
                "category_id": {
                    "type": "integer",
                    "example": 1
                },
                "cost_price": {
                    "type": "number",
                    "example": 12500
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer",
                    "example": 1
                },
                "description": {
                    "type": "string",
                    "example": "Gen 2 tv remote"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "product_attributes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProductAttributes"
                    }
                },
                "product_name": {
                    "type": "string",
                    "example": "TV Remote"
                },
                "quantity": {
                    "type": "integer",
                    "example": 30
                },
                "selling_price": {
                    "type": "number",
                    "example": 20000
                },
                "supplier_id": {
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "$ref": "#/definitions/models.NullTime"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        },
        "models.ProductAttributes": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "weight"
                },
                "product_id": {
                    "type": "integer",
                    "example": 1
                },
                "unit": {
                    "type": "string",
                    "example": "kg"
                },
                "value": {
                    "type": "string",
                    "example": "0.5"
                }
            }
        },
        "models.ProductAttributesRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "weight"
                },
                "unit": {
                    "type": "string",
                    "example": "kg"
                },
                "value": {
                    "type": "string",
                    "example": "0.5"
                }
            }
        },
        "models.ProductRequest": {
            "type": "object",
            "required": [
                "category_id",
                "cost_price",
                "product_name",
                "quantity"
            ],
            "properties": {
                "category_id": {
                    "type": "integer",
                    "example": 1
                },
                "cost_price": {
                    "type": "number",
                    "example": 12500
                },
                "description": {
                    "type": "string",
                    "example": "Gen 2 tv remote"
                },
                "product_attributes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProductAttributesRequest"
                    }
                },
                "product_name": {
                    "type": "string",
                    "example": "TV Remote"
                },
                "quantity": {
                    "type": "integer",
                    "example": 30
                },
                "selling_price": {
                    "type": "number",
                    "example": 20000
                },
                "supplier_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    }
}